# swagger.yml
...
swagger: "2.0"
info:
  title: Console Server
  version: 0.1.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
  - ws
basePath: /api/v1
...
securityDefinitions:
  key:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://auth.example.com/oauth/authorize
    tokenUrl: https://auth.example.com/oauth/token
  anonymous:
    name: X-Anonymous
    in: header
    type: apiKey
# Apply the key security definition to all APIs
security:
  - key: []
parameters:
  limit:
    name: limit
    in: query
    type: number
    format: int32
    default: 20
...
paths:
...
  /buckets/{bucket_name}/objects:
    get:
      summary: List Objects
      security:
        - key: []
        - anonymous: []
      operationId: ListObjects
      parameters:
        - name: bucket_name
          in: path
          required: true
          type: string
        - name: prefix
          in: query
          required: false
          type: string
        - name: recursive
          in: query
          required: false
          type: boolean
        - name: with_versions
          in: query
          required: false
          type: boolean
        - name: with_metadata
          in: query
          required: false
          type: boolean
        - $ref: "#/parameters/limit"
      responses:
        200:
          description: A successful response.
          schema:
            $ref: "#/definitions/listObjectsResponse"
        default:
          description: Generic error response.
          schema:
            $ref: "#/definitions/ApiError"
      tags:
        - Object
    delete:
      summary: Delete Object
      operationId: DeleteObject
      parameters:
        - name: bucket_name
          in: path
          required: true
          type: string
        - name: prefix
          in: query
          required: true
          type: string
        - name: version_id
          in: query
          required: false
          type: string
        - name: recursive
          in: query
          required: false
          type: boolean
        - name: all_versions
          in: query
          required: false
          type: boolean
        - name: non_current_versions
          in: query
          required: false
          type: boolean
        - name: bypass
          in: query
          required: false
          type: boolean
      responses:
        200:
          description: A successful response.
        default:
          description: Generic error response.
          schema:
            $ref: "#/definitions/ApiError"
      tags:
        - Object
...
  /kms/apis:
    get:
      summary: KMS apis
      operationId: KMSAPIs
      responses:
        200:
          description: A successful response.
          schema:
            $ref: "#/definitions/kmsAPIsResponse"
        default:
          description: Generic error response.
          schema:
            $ref: "#/definitions/ApiError"
      tags:
        - KMS
...
definitions:
...
  listObjectsResponse:
    type: object
    properties:
      objects:
        type: array
        items:
          $ref: "#/definitions/bucketObject"
        title: list of resulting objects
      total:
        type: integer
        format: int64
        title: number of objects
  bucketObject:
    type: object
    properties:
      name:
        type: string
      size:
        type: integer
        format: int64
        x-omitempty: false
      content_type:
        type: string
      last_modified:
        type: string
      is_latest:
        type: boolean
      is_delete_marker:
        type: boolean
      version_id:
        type: string
      user_tags:
        type: object
        additionalProperties:
          type: string
      expiration:
        type: string
      expiration_rule_id:
        type: string
      legal_hold_status:
        type: string
      retention_mode:
        type: string
      retention_until_date:
        type: string
      etag:
        type: string
      tags:
        type: object
        additionalProperties:
          type: string
      metadata:
        type: object
        additionalProperties:
          type: string
      user_metadata:
        type: object
        additionalProperties:
          type: string
...
  ApiError:
    type: object
    properties:
      message:
        type: string
      detailedMessage:
        type: string
...
  kmsAPIsResponse:
    type: object
    properties:
      results:
        type: array
        items:
          $ref: "#/definitions/kmsAPI"
  kmsAPI:
    type: object
    properties:
      method:
        type: string
      path:
        type: string
      maxBody:
        type: integer
      timeout:
        type: integer
...